requirements.txt
beautifulsoup4==4.11.1
Pillow==9.1.0
requests==2.27.1
tqdm==4.62.3
